+++ packages/data-loader/src/parsers/parseJson.ts
@@ -1,0 +1,5 @@
+import { KintoneRecordForParameter } from "../types";
+
+export const parseJson = (jsonString: string) => {
+  return JSON.parse(jsonString) as KintoneRecordForParameter[];
+};