+++ packages/data-loader/src/printers/printAsCsv/escapeDoubleQuotes.ts
@@ -1,0 +1,2 @@
+export const escapeDoubleQuotes = (fieldValue?: string | null) =>
+  fieldValue ? fieldValue.replace(/"/g, '""') : "";