+++ packages/plugin-packer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [4.0.6](https://github.com/kintone/js-sdk/compare/@kintone/plugin-packer@4.0.5...@kintone/plugin-packer@4.0.6) (2021-04-27)
+
+**Note:** Version bump only for package @kintone/plugin-packer
+
+
+
+
+
 ## [4.0.5](https://github.com/kintone/js-sdk/compare/@kintone/plugin-packer@4.0.4...@kintone/plugin-packer@4.0.5) (2021-04-20)
 
 **Note:** Version bump only for package @kintone/plugin-packer